Distinguish between shown and labeled in rank cutoff
