+++ src/js/kit/related-genes.js
@@ -29,6 +29,7 @@ import {
 
 import {writeLegend} from '../annotations/legend';
 import {getAnnotDomId} from '../annotations/process';
+import {applyRankCutoff} from '../annotations/labels';
 import {getDir} from '../lib';
 import initGeneCache from '../gene-cache';
 
@@ -480,37 +481,6 @@ function applyAnnotsIncludeList(annots, ideo) {
   return includedAnnots;
 }
 
-function applyRankCutoff(annots, ideo) {
-  if ('geneCache' in ideo === false) return annots;
-
-  const ranks = ideo.geneCache.interestingNames;
-
-  annots = annots.map(annot => {
-    annot.rank = ranks.indexOf(annot.name) || 1E10;
-    return annot;
-  });
-
-  // Ranks annots by popularity
-  const rankedAnnots = annots.sort((a, b) => {
-
-    // Search gene is most important, regardless of popularity
-    if (a.color === 'red') return -1;
-    if (b.color === 'red') return 1;
-
-    // Rank 3 is more important than rank 30
-    return a.rank - b.rank;
-  });
-
-  // console.log(rankedAnnots);
-  // console.log(annots.map(annot => annot.name));
-  // Take the 20 top-ranked genes
-  annots = rankedAnnots.slice(0, 20);
-
-  // console.log(annots.map(annot => {return annot.name + ' ' + annot.rank }));
-  // console.log(annots)
-  return annots;
-}
-
 /** Fetch and draw interacting genes, return Promise for annots */
 function processInteractions(annot, ideo) {
   return new Promise(async (resolve) => {
@@ -614,9 +584,9 @@ function finishPlotRelatedGenes(type, ideo) {
 
   annots = applyAnnotsIncludeList(annots, ideo);
   annots = mergeAnnots(annots);
-  annots = applyRankCutoff(annots, ideo);
+  annots = applyRankCutoff(annots, 40, ideo);
   ideo.relatedAnnots = mergeAnnots(annots);
-  ideo.relatedAnnots = applyRankCutoff(annots, ideo);
+  ideo.relatedAnnots = applyRankCutoff(annots, 40, ideo);
   annots.sort(sortAnnotsByRelatedStatus);
   ideo.relatedAnnots.sort(sortAnnotsByRelatedStatus);
